pyproject.toml
[project]
name = "gen-music"
version = "0.1.0"
description = "A modern CLI for generating music using Google's Vertex AI Lyria model."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pygame>=2.5.0",
    "rich>=13.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "requests>=2.32.5",
    "google-auth>=2.43.0",
    "pydub>=0.25.1",
]

[project.scripts]
gen-music = "gen_music.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gen_music"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
